requirements.txt
echo "discord.py" > requirements.txt
echo "aiohttp" >> requirements.txt
echo 'python-dotenv' > requirements.txt
discord-py
python-dotenv
flask
aiohttp
